adding game transactions controller

+++ CompetitiveGamingApp/Models/SingleGamePaymentTransactions.cs
@@ -6,6 +6,6 @@ public class SingleGamePaymentTransactions {
     public String? hostPlayer {get; set;}
     public String? gameId {get; set;}
     public String? playerLost {get; set;}
-    public int amountPaid {get; set;}
+    public double amountPaid {get; set;}
     public DateTime timePaid {get; set;}
 }
